backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
openai>=1.50.0
httpx>=0.27.0

# Phase 2: Multi-agent system
faiss-cpu>=1.7.0
numpy>=1.24.0
tiktoken>=0.5.0
